fix regression that broken git-based wraps with dirty contents
In commit 8da060706c2c34fbea91898509123cb190894f3e we fixed an issue
where stashing would ignore untracked files, potentially conflicting
with what we reset to.

Unfortunately in the process it broke the functionality entirely, by
producing an invalid git command... according to some versions of git.
This is actually a git bug, fixed upstream in git 2.37.0 via commit
b02fdbc80a41f73ceb6c99e8e27b22285243a335. It causes git itself to
re-synthesize the command line for a negation pathspec with a "."
rather than "", as that's what is needed to make it match all files
other than the negation.

For backwards compatibility to older versions of git, we manually pass
the dot pathspec ourselves.