backend/vs: ensure that build dir is preferred to src dir
Previously, the VS backend would add the the include directories
in the reverse order that the Ninja backend does it, which
means that the source directories would be preferred over
the build directories. For example, this would cause the
compiler to choose the file from the source directory
if a file with the same name is found in both.

Fixes 57ec097b5 ("vs: Use CompilerArgs() for compile and link args")
Fixes #11630