TMP: Cherry-pick of https://github.com/mesonbuild/meson/pull/11743
rust: Require `native: true` for `proc-macro` targets and allow linking into targets for other platforms

proc-macros are basically compiler extensions and must be compiled for
the build machine so that the compiler can load and execute it at
runtime.

Without this, cross-compilation of targets that use proc-macro
dependencies is going to fail.

Fixes https://github.com/mesonbuild/meson/issues/11702