dependencies: allow config-tool to fallback to default program names
If the dependency permits it, we can just do a PATH search instead of
mandating that it be listed in the cross file. This is useful for the
limited case where a specific dependency is known to be compatible with
any machine choice.

Mark the pybind11 dependency as supporting this. It's a valid choice
because pybind11 is a header-only C++ library.