Sort commands in the --help output semi-alphabetically
The list is now long enough that it becomes difficult to quickly spot a
command when it is unsorted. So let's create two alphabetically sorted
groups, the primary commands (compile/configure/dist/setup/install/test)
and the rest. This gives precedence to the commands most users are
looking for while still making it possible to find other commands.

Note that the help command is kept last intentionally to keep the
current functionality and make it stick out a bit.