modules/features: Add a module of functions for complex feature interactions
Feature options are a powerful tool in the Meson toolbox, especially for
dealing with dependencies. We provide a set of useful methods for
interactions between options, but there are a handful of cases we don't
handle, and don't make sense to add to features as methods.

One of the most requested features is the ability to generate feature
options within the meson.build environments. That's a trivial add for a
module.

The other big request we get is for allowing logical operations, `and`
and `or`. Which present a problem for features since no one can agree on
what `and` means. do `f1 and f2` mean `f1.allowed() and f2.allowed()`,
`f1.enabled() and f2.enabled()`,  `f1.value == f2.value()`, etc. the
same problem exists for `or`. This PR proposes a robust solution, using
`any` and `all` methods, which a mode keyword. This allows 2+ options to
be compared with a predicate that is user selected of either:

 auto : feat.is_auto()
 enabled : feat.is_enabled()
 disabled : feat.is_disabled()
 allowed : not feat.is_disabled()
 denied : not feat.is_enabled()

Which gives a user all of the possible combinations of options.