genvslite: improve readability of the mconf help text
The option description has a lot of data packed into it, which is
squeezed into a small, hard to read column. Give a bit less information,
focusing on essentials, to make it fit better.

Post-facto application of issues that were caught during post-merge
review, after the genvslite PR was merged without a full review.