avoid module indirection in name resolution for imported objects
We already import a bunch of objects directly from ..build but don't use
them nearly as much as we can. This resulted both in longer lines and s
minor performance difference since python has to resolve the name
binding the long way. There's no reason not to rewrite these names to
use the direct imports.

Found while investigating the fact that Executable was imported but
never used. It's easier to just use it.