linkers: fix mypy errors that were ignored due to lack of inheritance
A linker mixin has to be able to align with the base linker it will be
used for, in order to reference super(). Since they weren't inherited,
calls to super() resulted in mypy errors, which we ignored, and casting.

Use the same trick we use for compilers, and make the linker inherit
from the base linker type when running under mypy, and from object at
runtime.