Add 'install_name' property for targets
The 'install_name' property is used in place of 'name' when
constructing the install target path. This allows projects to generate
files using one name and then install them using another. This can be
used when generating multiple targets with the same basename but
different install directories within the same source directory.

Signed-off-by: Keith Packard <keithp@keithp.com>