build: Provide backward compatability for targets with path separators
To work around the restriction that only one target with a given name
could be created in any project directory, projects could prefix that
name with a unique directory name and things would "just work". Take
this example:

        static_library('foo/libbar',
                       [a.c, b.c],
                       install : true,
                       install_dir : 'foo'
                       )

        static_library('bletch/libbar',
                       [d.c, e.c],
                       install : true,
                       install_dir : 'bletch'
                       )

This would generate two libraries. The 'lib' prefix is prepended to
the directory names by the StaticLibrary code, while the target
names include the 'lib' prefix explicitly on 'libbar':

        libfoo/libbar.a
        libbletch/libbar.a.

Installation would copy the files. The copy code uses only the
basename from the source file when constructing the destination,
stripping off libfoo/ and libbletch/:

        libfoo/libbar.a         -> foo/libbar.a
        libbletch/libbar.a      -> bletch/libbar.a

We need to preserve this behavior while switching the internal
implementation to using the new 'install-name' support.

 1. When a target name includes a path separator, extract the basename
    and assign that to the 'install_name_auto' property. Then replace
    the path separators with '_' to 'flatten' the name, avoiding
    incidental creation of directories inside the build tree.

 2. When install_name is not set, use this install_name_auto instead,
    first removing any matching prefix (e.g. 'lib' for libraries) as
    those will be re-added when the install filename is created.

Projects which depend on this behavior can be converted to use
'install-name' directly, once they depend on a version of meson with
that feature.

Signed-off-by: Keith Packard <keithp@keithp.com>