mtest: try a bit harder to avoid weird non-parseable output at startup
In commit 628effb3698e85e403351d2705d576cf4ee8c50c we started verifying
the build.ninja file was up to date before even `--list`ing tests,
because we could end up with incorrect information. This meant that
ninja always runs at startup, and typically returns "no work to do",
which ended up listed as "one of" the tests.

Instead of unconditionally running ninja attached to the console, first
check it in dry-run mode with stdout intercepted, to see if ninja
considers itself up to date. If it is, continue. Only if an actual
refresh is needed, do we run it while attached to the console.

In the garden path, this avoids useless information. In cases where we'd
already print a full meson reconfigure log, we continue to do so.