Always apply the default_options
In the implementation of `project()`, the `default_options` are only
applied on the first invocation and not on later ones. This seems to
have been a workaround to avoid disturbing the command line options on
reconfigure (which weren't actually present, so...? Confusing.)

Make this application unconditional so any options set by
`default_options` are still set after a reconfigure.

Note that this means changes to `default_options` *will* take effect
after a reconfiguration (if not previously overridden on the command
line). This is a user-visible change in behavior from 1.1.1 and prior.