xcode: Set bridging header for mixed swift targets
This must be explicitly set in Xcode or it will not be able to compile
mixed targets successfully. This is not needed for pure Swift targets.

Set a new variable "is_swift" so finding the bridging header does not
take O(n^2) time.

Fixes passing swift test case 5 (mixed) with Xcode backend.