CustomTarget: Do not recurse in get_transitive_build_target_deps()
The goal of that function is to collect BuildTarget this custom target
uses. E.g. if there are dll or exe passed as argument to a python
script, chances are that we need to set PATH to be able to use them.

The recursion happens already when we call get_all_link_deps() on those
BuildTarget.

Also fix its usage in test(), which was weirdly using it only if the
command first argument is a CustomTarget, but not for for other
arguments. The intended usage was to simply call
determine_windows_extra_paths() with all build targets from all command
arguments, like the backend does for custom target's command.