allow some ObjectHolder subclasses to continue to be generic
ExternalProgram and CustomTarget have some use cases for producing
subclassed interpreter holders with more specific types and methods. In
order for those subclasses to properly refer to their held_object, we
need a shared base class that is still generic, though bound.

For the derived held objects, inherit from the base class and specify
the final types as the module-specific type.